requirements.txt
google-auth-oauthlib>=0.4.6
google-auth>=2.6.0
requests>=2.27.1
selenium>=4.1.0
webdriver-manager>=3.5.2
python-dotenv>=0.19.2